+++ src/main/java/ladder/domain/ladders/RandomPointPredicate.java
@@ -1,4 +1,4 @@
-package ladder.domain.ladder;
+package ladder.domain.ladders;
 
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
